Overlay purelib contents correctly
In tensorflow-1.15.0 wheel, there is a tensorflow_core directory both at
the root of the wheel, as well as under the purelib directory.  When
extracting the wheel, piptool was using shutil.move to move the
purelib/tensorflow_core to the wheel root, but it would end up in
tensorflow_core/tensorflow_core.  Use shutil.copy2 and shutil.rmtree to
perform a "merging and overwriting" move instead.